+++ L-CAD/lcadengine/src/utils/fastlist.h
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
-
+#include "stddef.h"
 
 #ifndef FASTLIST_H_INCLUDED
 #define FASTLIST_H_INCLUDED
@@ -31,7 +31,7 @@ typedef struct s_fastlist fastlist;
  * @returns fastlist            Pointer to a new fastlist
  *                              FASTLIST_FAILED on error
  */
-fastlist *fastlistCreate(unsigned long size);
+fastlist *fastlistCreate(size_t size);
 
 
 /**
@@ -47,12 +47,12 @@ void fastlistDelete(fastlist *ctx);
  * @brief Get's <index> element from list 
  *
  * @param fastlist *ctx         An initialised fastlist to work on
- * @param unsigned long index   The index of the item
+ * @param size_t index   The index of the item
  *
  * @returns void*               Item from the fastlist
  *                              FASTLIST_FAILED on error
  */
-void *fastlistGetIndex(fastlist *ctx, unsigned long index);
+void *fastlistGetIndex(fastlist *ctx, size_t index);
 
 
 /**
@@ -62,7 +62,7 @@ void *fastlistGetIndex(fastlist *ctx, unsigned long index);
  *
  * @returns fastlist            Size of the fastlist
  */
-unsigned long fastlistSize(fastlist *ctx);
+size_t fastlistSize(fastlist *ctx);
 
 
 /**
@@ -71,21 +71,21 @@ unsigned long fastlistSize(fastlist *ctx);
  * @param fastlist *ctx         An initialised fastlist to work on
  * @param void *item            Pointer to an item to be added to the list
  *
- * @returns unsigned long       New length of the list
+ * @returns size_t       New length of the list
  */
-unsigned long fastlistAdd(fastlist *ctx, void *item);
+size_t fastlistAdd(fastlist *ctx, void *item);
 
 
 /**
  * @brief Removes item of index from the list and returns it (pop)
  *
  * @param fastlist *ctx         An initialised fastlist to work on
- * @param unsigned long index   The index of the item in the list
+ * @param size_t index   The index of the item in the list
  *
  * @returns void*               Item removed from the list (aka POP)
  *                              FASTLIST_FAILED on error
  */
-void *fastlistRemoveIndex(fastlist *ctx, unsigned long index);
+void *fastlistRemoveIndex(fastlist *ctx, size_t index);
 
 
 /**
@@ -94,21 +94,21 @@ void *fastlistRemoveIndex(fastlist *ctx, unsigned long index);
  * @param fastlist *ctx         An initialised fastlist to work on
  * @param void *pointer         Pointer to an object in the fastlist
  *
- * @returns unsigned long       Index removed from the fastlist
+ * @returns size_t       Index removed from the fastlist
  *                              FASTLIST_FAILED on error
  */
-unsigned long fastlistRemoveByPointer(fastlist *ctx, void* pointer);
+size_t fastlistRemoveByPointer(fastlist *ctx, void* pointer);
 
 
 /**
  * @brief NOT IMPLEMENTED
  *
  * @param fastlist *ctx         An initialised fastlist to work on
- * @param unsigned long size    Some size
+ * @param size_t size    Some size
  *
  * @returns void **             Direct pointers to the fastlist
  *                              FASTLIST_FAILED on error
  */
-void **fastlistDMA(fastlist *ctx, unsigned long size);
+void **fastlistDMA(fastlist *ctx, size_t size);
 
 #endif
